pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "cairosvg>=2.7.1",
    "cryptography>=44.0.2",
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "numpy>=2.2.4",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "pymysql>=1.1.1",
    "python-dotenv>=1.1.0",
    "routes>=2.5.1",
    "sqlalchemy>=2.0.40",
    "werkzeug>=3.1.3",
]
